Plotly graphs as static img for viewing on github
Using kaleido version 0.2.0 because latest version 0.2.1 is throwing  while performing The following packages are already present in the pyproject.toml and will be skipped:

  • kaleido

If you want to update it to the latest compatible version, you can use `poetry update package`.
If you prefer to upgrade it to the latest available version, you can use `poetry add package@latest`.

Nothing to add.